3ºano/PECI/ProjetoFaceRecognition/CRUD_Albuns_BE/requirements.txt
fastapi[all]==0.68.1
supabase-py==0.3.2